+++ src/core/core/tool_api.cpp
@@ -37,7 +37,7 @@
 #include <dsn/utility/factory_store.h>
 #include <dsn/utility/singleton_store.h>
 #include "service_engine.h"
-#include "message_parser_manager.h"
+#include "core/rpc/message_parser_manager.h"
 
 namespace dsn {
