+++ src/core/rpc/network.sim.h
