+++ src/core/tools/common/providers.common.cpp
@@ -33,15 +33,15 @@
  *     xxxx-xx-xx, author, fix bug about xxx
  */
 
-#include "asio_net_provider.h"
+#include "core/rpc/asio_net_provider.h"
 #include <dsn/tool/providers.common.h>
 #include "lockp.std.h"
 #include "simple_task_queue.h"
-#include "network.sim.h"
+#include "core/rpc/network.sim.h"
 #include "simple_logger.h"
-#include "dsn_message_parser.h"
-#include "thrift_message_parser.h"
-#include "raw_message_parser.h"
+#include "core/rpc/dsn_message_parser.h"
+#include "core/rpc/thrift_message_parser.h"
+#include "core/rpc/raw_message_parser.h"
 
 namespace dsn {
 namespace tools {
